tmq administrator client is added
