+++ src/Twino.Client.TMQ/Connectors/TmqStickyConnector.cs
@@ -43,9 +43,7 @@ namespace Twino.Client.TMQ.Connectors
             _reader = new MessageReader(serailizationAction);
         }
 
-        /// <summary>
-        /// Called when a message is received from server
-        /// </summary>
+        /// <inheritdoc />
         protected override void ClientMessageReceived(ClientSocketBase<TmqMessage> client, TmqMessage payload)
         {
             base.ClientMessageReceived(client, payload);
